Refactor CloudKit protocols and move implementations
Merged CKInitializable and CKRepresentable protocols into CKPersistable in a new file. Moved protocol conformances for EventObject and SessionObject to CKPersistable.swift. Relocated Matrix's CKInitializable implementation from Matrix+CKInitializable.swift to Matrix.swift, and removed the now redundant protocol files.
