+++ Sources/Scout/Core/Matrix.swift
@@ -43,6 +43,53 @@ extension Matrix: Combining {
     }
 }
 
+extension Matrix: CKInitializable {
+    enum MapError: LocalizedError {
+        case missingDate
+        case missingName
+        case missingCells
+        case invalidCells
+
+        var errorDescription: String? {
+            switch self {
+            case .missingDate:
+                "Missing date field"
+            case .missingName:
+                "Missing name field"
+            case .missingCells:
+                "Missing cells"
+            case .invalidCells:
+                "Invalid cells. Expected a dictionary of Strings to Int or Double"
+            }
+        }
+    }
+
+    init(record: CKRecord) throws {
+        guard let date = record["date"] as? Date else {
+            throw MapError.missingDate
+        }
+        guard let name = record["name"] as? String else {
+            throw MapError.missingName
+        }
+
+        self.date = date
+        self.name = name
+        self.recordID = record.recordID
+
+        let cellKeys = record.allKeys().filter { $0.hasPrefix("cell_") }
+        let cellDict = record.dictionaryWithValues(forKeys: cellKeys)
+
+        guard cellDict.count > 0 else {
+            throw MapError.missingCells
+        }
+        guard let cellDict = cellDict as? [String: T.Value] else {
+            throw MapError.invalidCells
+        }
+
+        self.cells = try cellDict.map(T.init)
+    }
+}
+
 extension Matrix: CustomStringConvertible {
     var description: String {
         "Matrix(\(name), \(date), \(cells.count) cells)"
